use pointers for slog handlers
